Updated 404 page, fixed css, seperate posts into regular and projects, added tempory project

+++ _includes/typewriter.js
@@ -1,29 +1,39 @@
 <script>
-var title = document.getElementById('title');
+var title = document.getElementById('page-title');
 
 var siteName = {{ page.title | jsonify }};
 
 if (siteName == null)
 {
-    var typewriter = new Typewriter(title, {
-        loop: true});
+    if ({{ page.path | jsonify }} == "404.html")
+    {
+        var typewriter = new Typewriter(title, {
+            loop: false});
 
-    typewriter.typeString('Hello world!')
-        .pauseFor(2500)
-        .deleteAll()
-        .typeString("I'm a game developer")
-        .pauseFor(2500)
-        .deleteChars(14)
-        .typeString('web developer')
-        .pauseFor(2500)
-        .deleteChars(13)
-        .typeString('linux tinkerer')
-        .pauseFor(2500)
-        .deleteChars(14)
-        .typeString('problem solver')
-        .pauseFor(2500)
-        .deleteAll()
-        .start();
+        typewriter.typeString("/404").start();
+    }
+    else
+    {
+        var typewriter = new Typewriter(title, {
+            loop: true});
+
+        typewriter.typeString("Hello world!")
+            .pauseFor(2500)
+            .deleteAll()
+            .typeString("I'm a game developer")
+            .pauseFor(2500)
+            .deleteChars(14)
+            .typeString('web developer')
+            .pauseFor(2500)
+            .deleteChars(13)
+            .typeString('linux tinkerer')
+            .pauseFor(2500)
+            .deleteChars(14)
+            .typeString('problem solver')
+            .pauseFor(2500)
+            .deleteAll()
+            .start();
+    }
 }
 else
 {
@@ -32,12 +42,13 @@ else
 
     var directory = '/';
 
-    {% if page.path contains '_posts' %}
-        {% unless page.path contains '_pages' %}
-            directory += "Posts/";
-        {% endunless %}
+    {% if page.path contains '_posts/posts/' %}
+        directory += "Posts/";
+    {% elsif page.path contains '_posts/projects/' %}
+        directory += "Projects/";
     {% endif %}
 
+    // Enclose in quotes if string has  whitespace
     if (/\s/g.test(siteName))
         siteName = '\'' + siteName + '\'';
